refactor(extensions): move CacheStore + TokenCacheStore to src/extensions/cache/, rename ext-redis → ext-cache-redis
- src/extensions/cache/{cache-store,token-cache-store,index}.ts (new layout)
- src/extensions/interfaces/index.ts shim re-exports from ../cache/
- extensions/ext-cache-redis (@veryfront/ext-cache-redis)
- Root deno.json workspace + ./extensions/cache export
- Consumer stragglers (recommendations.ts, runtime name, JSDoc, tests) follow Task 2/3 precedent
