+++ extensions/ext-cache-redis/deno.json
@@ -1,5 +1,5 @@
 {
-  "name": "@veryfront/ext-redis",
+  "name": "@veryfront/ext-cache-redis",
   "version": "0.1.0",
   "exports": "./src/index.ts",
   "veryfront": {
