+++ src/extensions/cache/index.ts
@@ -1,0 +1,12 @@
+/**
+ * Cache category barrel — generic cache and proxy-grade token cache.
+ *
+ * @module extensions/cache
+ */
+
+export type { CacheStore } from "./cache-store.ts";
+export type {
+  TokenCacheEntry,
+  TokenCacheStats,
+  TokenCacheStore,
+} from "./token-cache-store.ts";
